pyproject.toml
[project]
name = "ai-hobby-labe"
version = "0.1.0"
description = "A collection of simple AI-based scripts for personal automations"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.1",
    "yt-dlp>=2025.1.15",
    "openai>=1.59.7",
    "brave-search>=0.1.8",
    "requests>=2.32.3",
    "llama-index>=0.12.12",
    "llama-index-llms-openai-like>=0.3.3",
    "markdown-pdf>=1.3.3",
    "markdown2>=2.5.3",
    "pillow>=11.1.0",
]
